tidy: Fix a couple of "narrowing conversions" warning.
Tidy is warning about a couple of uints that are being assigned to
int.  This appears to be a result of an enum size being changed from
uint->uint8_t, and a function that takes that enum as an argument not
also having its size changed.

These were introduced in bbe421d008.

+++ mythtv/libs/libmythtv/dtvconfparserhelpers.h
@@ -696,10 +696,10 @@ class DTVModulationSystem : public DTVParamHelper
                   "Modulation System types don't match DVB includes.");
 #endif
 
-    explicit DTVModulationSystem(uint _value = kModulationSystem_UNDEFINED)
+    explicit DTVModulationSystem(uint8_t _value = kModulationSystem_UNDEFINED)
         : DTVParamHelper(_value) { }
 
-    DTVModulationSystem& operator=(uint _value)
+    DTVModulationSystem& operator=(uint8_t _value)
         { m_value = _value; return *this; }
 
     bool IsCompatible(const DTVModulationSystem other) const
